Apply fixes to all inline elements
